rename: PostContainer로 컴포넌트명 변경 및 파일 이동(Next-Study-habjung#56)
