+++ src/components/pages/ProfilePage.tsx
@@ -2,7 +2,7 @@
 
 import { postListData } from '@/__mock__/responseData/postList';
 import { productList } from '@/__mock__/responseData/productList';
-import PostSection from '@/components/profile/post/PostSection';
+import PostSection from '@/components/profile/container/PostContainer';
 import ProductListContainer from '@/components/profile/container/ProductListContainer';
 import UserInfoContainer from '@/components/profile/container/UserInfoContainer';
 
@@ -29,13 +29,13 @@ export default function ProfilePage({ accountName }: IProfilePageProps) {
     <main className="flex flex-col gap-6px bg-gray-50">
       <UserInfoContainer
         userProfile={userProfile}
-        className="bg-white pb-26px pt-30px shadow-test1"
+        className="bg-white pb-26px pt-30px"
       />
       <ProductListContainer
-        className="bg-white py-5 shadow-test2"
+        className="bg-white py-5"
         productList={productList}
       />
-      <PostSection className="bg-white shadow-test4" postList={postListData} />
+      <PostSection className="bg-white" postList={postListData} />
     </main>
   );
 }
