+++ src/components/profile/container/PostContainer.tsx
@@ -5,24 +5,26 @@ import { useState } from 'react';
 
 import { IPostList } from '@/api/types/post';
 
-import { IconSwitcher } from './IconSwitcher';
-import { PostAlbumView, PostListView } from './PostView';
+import { IconSwitcher } from '../components/post/IconSwitcher';
+import { PostAlbumView, PostListView } from '../components/post/PostView';
 
 interface IPostSection {
   className?: string;
   postList: IPostList;
 }
 
-export default function PostSection({ className, postList }: IPostSection) {
+export default function PostContainer({ className, postList }: IPostSection) {
   const [postType, setPostType] = useState<'list' | 'album'>('list');
+
+  const handlePostType = (type: 'list' | 'album') => {
+    if (type === postType) return;
+    console.log('change일 때만');
+    setPostType(type);
+  };
 
   return (
     <section className={classNames('', className)}>
-      <IconSwitcher
-        postType={postType}
-        onListClick={() => setPostType('list')}
-        onAlbumClick={() => setPostType('album')}
-      />
+      <IconSwitcher postType={postType} onSwitch={handlePostType} />
       {postType === 'list' && <PostListView postList={postList} />}
       {postType === 'album' && <PostAlbumView postList={postList} />}
     </section>
